Rholang: Move files around to better follow scala style.
The compiler libraries are left alone for the moment, but the generated java
sources are now under coop.rchain.rholang for consistency.
